% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{avg_sliding_window}
\alias{avg_sliding_window}
\title{Sliding window convolution.}
\usage{
avg_sliding_window(x, window)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Integer vector for sliding window.}
}
\value{
Numeric vector of averaged values over the sliding window.
}
\description{
Sliding window convolution.
}
