% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{inv_fft}
\alias{inv_fft}
\title{Inverse Fast Fourier Transform}
\usage{
inv_fft(x, len = NULL)
}
\arguments{
\item{x}{Numeric vector.}

\item{len}{Optional argument, define length of provided vector as a numeric.}
}
\value{
Numeric vector.
}
\description{
Inverse Fast Fourier Transform
}
