% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliencymap.R
\name{saliency_map}
\alias{saliency_map}
\title{Create saliency map}
\usage{
saliency_map(x, window = 3)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Positive integer value.}
}
\value{
Numeric vector representing the saliency map values.
}
\description{
Create saliency map
}
\examples{
tmp <- ts(rnorm(12*6,10,2), start=c(2009, 1), end=c(2014, 12), frequency=12)

}
