% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extrapolate.R
\name{simple_extrapolation}
\alias{simple_extrapolation}
\title{Simple extrapolation of time series}
\usage{
simple_extrapolation(x, m = 5)
}
\arguments{
\item{x}{Numeric vector of data points to extrapolate forwards.}

\item{m}{Integer value representing the number of preceding points for the estimation.}
}
\value{
Vector of estimated points.
}
\description{
The next point is estimated based upon a user provided number (m) of preceeding points. The gradient
between the last point and the prior points are summed and applied to the next point based upon the n-m+1 data point.
}
\examples{
tmp <- ts(rnorm(12*6,10,2), start=c(2009, 1), end=c(2014, 12), frequency=12)
simple_extrapolation(tmp, 5)
simple_extrapolation(tmp, 2)
simple_extrapolation(tmp, 10)
}
