% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{add_anomaly}
\alias{add_anomaly}
\title{Generate anomalies within time series}
\usage{
add_anomaly(ts, n, score_window, spec_window)
}
\arguments{
\item{n}{Integer value for number of anomalies to apply}

\item{score_window}{Integer value for the window width for scoring.}

\item{spec_window}{Positive integer value for the window to calculate the averaged log spectrum.}
}
\description{
Add \code{n} anomalies to an existing vector.
}
\details{
Method to calculate is:

\deqn{x + (\bar{x} + mean)(1+var)(\mathcal{N}(0,1))}

Will exclude the first point to ensure variance
calculation has a useful value.
}
