% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/saliencymap.R
\name{saliency_map}
\alias{saliency_map}
\title{Create saliency map}
\usage{
saliency_map(x, window = 3)
}
\arguments{
\item{x}{Numeric vector.}

\item{window}{Positive integer value.}
}
\value{
Numeric vector representing the saliency map values.
}
\description{
Using the provided numeric input, typically a time series, calculate the spectral residual
and output the saliency map for use in anomaly detection.
}
\examples{
tmp <- ts(rnorm(12*6,10,2), start=c(2009, 1), end=c(2014, 12), frequency=12)
saliency_map(tmp)
}
\seealso{
\code{\link{anomaly_score}}
}
