% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralAnomaly-package.R
\docType{package}
\name{spectralAnomaly-package}
\alias{spectralAnomaly}
\alias{spectralAnomaly-package}
\title{spectralAnomaly: Detect Anomalies Using the 'Spectral Residual' algorithm}
\description{
Apply the 'Spectral Residual' algorithm to data, such as a time series, to detect anomalies. Anomaly scores can be used determine outliers based upon a threshold or fed into more sophistcated prediction models.
}
\author{
\strong{Maintainer}: Allen OBrien \email{allen.g.obrien@gmail.com} [copyright holder]

}
\keyword{internal}
