% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralAnomaly-package.R
\docType{package}
\name{spectralAnomaly-package}
\alias{spectralAnomaly}
\alias{spectralAnomaly-package}
\title{spectralAnomaly: Detect Anomalies Using the 'Spectral Residual' Algorithm}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Apply the spectral residual algorithm to data, such as a time series, to detect anomalies. Anomaly scores can be used to determine outliers based upon a threshold or fed into more sophisticated prediction models. Methods are based upon "Time-Series Anomaly Detection Service at Microsoft", Ren, H., Xu, B., Wang, Y., et al., (2019) \doi{10.48550/arXiv.1906.03821}.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://al-obrien.github.io/spectralAnomaly/}
  \item \url{https://github.com/al-obrien/spectralAnomaly}
  \item Report bugs at \url{https://github.com/al-obrien/spectralAnomaly/issues}
}

}
\author{
\strong{Maintainer}: Allen OBrien \email{allen.g.obrien@gmail.com} [copyright holder]

}
\keyword{internal}
